requirements.txt
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.9.0
attrs==23.2.0
autopep8==2.3.1
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.8
dateparser==1.2.1
decorator==5.1.1
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et_xmlfile==2.0.0
fake-useragent==2.0.3
filelock==3.17.0
flashtext==2.7
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
jdatetime==4.1.0
pywin32==306
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.42
googletrans==4.0.2
greenlet==3.0.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaitaistruct==0.10
kiwisolver==1.4.8
langcodes==3.3.0
langdetect==1.0.9
logger==1.4
logging==0.4.9.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
patsy==0.5.6
pdfminer3==2018.12.3.0
phonenumbers==9.0.3
pillow==10.2.0
plotly==5.19.0
plotly-express==0.4.1
preshed==3.0.9
protobuf==4.25.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyaes==1.6.1
pyarmor==7.6.1
pyarmor.cli.core==7.6.2
pyarrow==15.0.0
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
pydivert==2.1.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-instagram==1.3.2
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
ratelim==0.1.6
referencing==0.33.0
regex==2023.12.25
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.5.3
scipy==1.12.0
seaborn==0.13.2
selenium==4.29.0
selenium-stealth==1.0.6
selenium-wire==5.1.0
sentencepiece==0.2.0
setuptools==69.1.1
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.4.8
statsmodels==0.14.1
streamlit==1.31.1
streamlit-tags==1.2.8
sympy==1.13.1
Telethon==1.39.0
tenacity==8.2.3
thinc==8.2.3
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
toolz==0.12.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.66.2
transformers==4.51.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.3.85
ultralytics-thop==2.0.14
undetected-chromedriver==3.5.5
urllib3==2.3.0
validators==0.22.0
wasabi==1.1.2
watchdog==4.0.0
weasel==0.3.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
zstandard==0.23.0
